+++ src/Order/resources/config/module.php
@@ -9,16 +9,17 @@ return [
             'start_sequence_from' => 1,
             'prefix' => '',
             'pad_length' => 1,
-            'pad_string' => '0'
+            'pad_string' => '0',
         ],
         'time_hash' => [
             'high_variance' => false,
             'start_base_date' => '2000-01-01',
-            'uppercase' => false
+            'uppercase' => true,
+            'extra_digit' => false,
         ],
         'nano_id' => [
             'size' => 12,
-            'alphabet' => '0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZ'
+            'alphabet' => '0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZ',
         ],
     ]
 ];